+++ packages/@storefront/navigation/src/filter-refinement-controls/index.ts
@@ -1,4 +1,5 @@
-import { tag } from '@storefront/core';
+import { tag, StoreSections, Tag } from '@storefront/core';
+import RefinementControls from '../refinement-controls';
 import ValueRefinementControls from '../value-refinement-controls';
 
 @tag('gb-filter-refinement-controls', require('./index.html'))
@@ -7,10 +8,23 @@ class FilterRefinementControls extends ValueRefinementControls {
     return 'filterControls';
   }
 
+  get selectMultipleRefinements(): string {
+    switch (this.props.storeSection) {
+      case StoreSections.PAST_PURCHASES:
+        return 'selectMultiplePastPurchaseRefinements';
+      case StoreSections.SEARCH:
+        return 'selectMultipleRefinements';
+    }
+  }
+
   fetchMoreRefinements = () => {
     if (this.state.more) {
       this.state.moreRefinements();
     }
+  }
+
+  selectMatchedRefinements = (event: Tag.Event, refinements: RefinementControls.SelectedRefinement[]) => {
+    this.actions[this.selectMultipleRefinements](this.props.navigation.field, refinements.map((refinement) => refinement.index));
   }
 }