+++ packages/@storefront/structure/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] [minor]
+### Added
+- SF-1073: Update `gb-filtered-list` component to have a "Select All" button.
+  - `gb-filtered-list` now has a "Select All" button that will select all of the available refinements after an initial filter.
+
 ## [1.47.1] - 2019-01-14
 ### Changed
 - SF-1241: Update filtered list component to use table of DOM key strings.